refactor the bot/ update README
