+++ cogs/__init__.py
@@ -1,0 +1,1 @@
+# cogs/__init__.py
