+++ utils/chroma_client.py
@@ -1,0 +1,5 @@
+import chromadb
+
+chroma_client = chromadb.PersistentClient(path="./chroma_db")
+
+collection = chroma_client.get_or_create_collection("personal_notes")
